requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.7.1
python-dotenv==1.1.0
anthropic==0.51.0
pyodbc==5.2.0
sqlalchemy==2.0.40
pytest==8.3.5
requests==2.32.3
httpx==0.28.1
azure-identity==1.21.0
